+++ src/view/filters.js
@@ -1,11 +1,11 @@
-export const filtersTemplate = () => (
-  `
-    <form class="trip-filters" action="#" method="get">
+import { createElement } from '../render';
+
+const createFiltersMarkup = () => (
+  `<form class="trip-filters" action="#" method="get">
         <div class="trip-filters__filter">
             <input id="filter-everything" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="everything" checked>
             <label class="trip-filters__filter-label" for="filter-everything">Everything</label>
         </div>
-
         <div class="trip-filters__filter">
             <input id="filter-future" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="future">
             <label class="trip-filters__filter-label" for="filter-future">Future</label>
@@ -15,8 +15,25 @@ export const filtersTemplate = () => (
             <input id="filter-past" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="past">
             <label class="trip-filters__filter-label" for="filter-past">Past</label>
         </div>
-
         <button class="visually-hidden" type="submit">Accept filter</button>
-    </form>
-  `
+    </form>`
 );
+
+export default class filters {
+  #element = null;
+
+  get element() {
+    if (!this.#element) {
+      this.#element = createElement(this.template);
+    }
+    return this.#element;
+  }
+
+  get template() {
+    return createFiltersMarkup();
+  }
+
+  removeElement() {
+    this.#element = null;
+  }
+}
