--- src/view/navigationTemplate.js
@@ -1,8 +1,0 @@
-export const navigationTemplate = () => (
-  `
-    <nav class="trip-controls__trip-tabs  trip-tabs">
-        <a class="trip-tabs__btn  trip-tabs__btn--active" href="#">Table</a>
-        <a class="trip-tabs__btn" href="#">Stats</a>
-    </nav>
-  `
-);
