+++ src/view/sort.js
@@ -1,6 +1,6 @@
-export const sortTemplate = () => (
-  `
-    <form class="trip-events__trip-sort  trip-sort" action="#" method="get">
+import { createElement } from '../render';
+export const createSortMarkup = () => (
+  `<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
             <div class="trip-sort__item  trip-sort__item--day">
               <input id="sort-day" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-day" checked>
               <label class="trip-sort__btn" for="sort-day">Day</label>
@@ -25,6 +25,25 @@ export const sortTemplate = () => (
               <input id="sort-offer" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-offer" disabled>
               <label class="trip-sort__btn" for="sort-offer">Offers</label>
             </div>
-          </form>
-  `
+          </form>`
 );
+
+
+export default class sort {
+  #element = null;
+
+  get element() {
+    if (!this.#element) {
+      this.#element = createElement(this.template);
+    }
+    return this.#element;
+  }
+
+  get template() {
+    return createSortMarkup();
+  }
+
+  removeElement() {
+    this.#element = null;
+  }
+}
